Moved admin page into blueprint
