+++ app/errors.py
@@ -1,5 +1,6 @@
 from flask import render_template
-from app import app, db
+from app import app
+from app.models import db
 
 @app.errorhandler(404)
 def not_found_error(error):
