+++ app/forms.py
@@ -1,7 +1,9 @@
+from flask import current_app
 from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField, PasswordField, BooleanField, SelectField
 from wtforms.validators import DataRequired, EqualTo, Email, ValidationError
-from app.models import User, Role
+from app import app
+from app.models import User, Role, db
 
 class RegistrationForm(FlaskForm):
   f_name = StringField('First Name', validators=[DataRequired()])
@@ -25,12 +27,13 @@ class LoginForm(FlaskForm):
 
 
 class AdminForm(FlaskForm):
-  f_name = StringField('First Name', validators=[DataRequired()])
-  l_name = StringField('Last Name', validators=[DataRequired()])
-  email = StringField('Email', validators=[DataRequired(), Email()])
-  role = SelectField('Role', validators=[DataRequired()], choices=[(i.id, i.name) for i in Role.query.all()], coerce=int)
-  # role = StringField('Role', validators=[DataRequired()])
-  submit = SubmitField('Create User')
+  with app.app_context():
+    f_name = StringField('First Name', validators=[DataRequired()])
+    l_name = StringField('Last Name', validators=[DataRequired()])
+    email = StringField('Email', validators=[DataRequired(), Email()])
+    role = SelectField('Role', validators=[DataRequired()], choices=[(i.id, i.name) for i in Role.query.all()], coerce=int)
+    # role = StringField('Role', validators=[DataRequired()])
+    submit = SubmitField('Create User')
 
   def validate_email(self, email):
     u = User.query.filter_by(email=email.data).first()
