+++ app/models.py
@@ -1,8 +1,12 @@
-from app import db, login
+
+from app import login
 from datetime import datetime
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
 
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
 
 class User(UserMixin, db.Model):
   id = db.Column(db.Integer, primary_key=True)
