+++ app/routes.py
@@ -1,8 +1,8 @@
-from app import app, db
-from flask import redirect, render_template, url_for, flash, request
+from app import app
+from flask import redirect, render_template, url_for, flash, request, Blueprint
 from flask_login import current_user, login_required, login_user, logout_user
 from app.forms import LoginForm, RegistrationForm, AdminForm
-from app.models import User
+from app.models import User, db
 
 @app.route('/')
 def index():
@@ -62,31 +62,3 @@ def logout():
   logout_user()
   flash('You are logged out.', 'success')
   return redirect(url_for('login'))
-
-
-@login_required
-@app.route('/admin', methods=['GET', 'POST'])
-def admin():
-  if not current_user.is_authenticated or not current_user.role.name == 'Super User':
-    flash('You are not authorized to view this page.', 'danger')
-    return redirect(url_for('index'))
-  form = AdminForm()
-  if form.validate_on_submit():
-    print(form.role.data)
-    print("begin validation...")
-    user = User.query.filter_by(email=form.email.data).first()
-    if user is not None:
-      flash('That email is already taken. Choose another.', 'danger')
-      return redirect(url_for('admin'))
-    new_user = User(f_name=form.f_name.data, l_name=form.l_name.data, email=form.email.data, role_id=form.role.data)
-    new_user.set_password(form.email.data)
-    db.session.add(new_user)
-    db.session.commit()
-    flash('New user created.', 'success')
-    return redirect(url_for('admin'))
-  context = {
-    'form': AdminForm(),
-    'description': 'Create a new user',
-    'title': 'Admin'
-  }
-  return render_template('admin/admin.html', **context)
